+++ Ranger/Services/SharkScope/Metadata/Metadata.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 
-struct Response: Decodable
+struct Metadata: Decodable
 {