+++ Ranger/Services/SharkScope/Metadata/MetadataRequest.swift
@@ -1,0 +1,21 @@
+//
+//  LiveTourneyTableQuery.swift
+//  Ranger
+//
+//  Created by Geri Borbás on 2019. 12. 14..
+//  Copyright © 2019. Geri Borbás. All rights reserved.
+//
+
+import Foundation
+
+
+struct MetadataRequest: Request
+{
+    
+    
+    typealias ResponseType = Metadata
+    
+    
+    var path: String { "metadata" }
+    var parameters:  [String: String] { [:] }
+}